+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 899
+/// Strings: 900
 ///
-/// Built on 2026-06-19 at 15:52 UTC
+/// Built on 2026-06-20 at 12:35 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import